+++ History.md
@@ -4,6 +4,7 @@
 
 - An optional mechanism for background job execution is introduced, making it possible to send e-mails asynchronously (therefore not blocking regular requests).
 - The deadline circle at the top of the homepage now also shows dates if more than one future deadline exists. It can also be deactivated entirely.
+- The export functions for motions in the admin list now allow to combine one or multiple motion types, with all motion types being selected by default.
 
 ### Version 4.15.3 [not released yet]